Refactor: #17 Devide file
- LocalPlayerViewController를 별개의 파일로 나눔
