+++ PiPPl/View/LocalVideo/LocalVideoGalleryViewController.swift
@@ -201,31 +201,6 @@ extension LocalVideoGalleryViewController: UICollectionViewDelegate {
     }
 }
 
-class LocalPlayerViewController: AVPlayerViewController {
-
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-        self.removeFromParent()
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        tabBarController?.tabBar.isHidden = true
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        player?.play()
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        player?.pause()
-        tabBarController?.tabBar.isHidden = false
-    }
-
-}
-
 extension LocalVideoGalleryViewController: AVPlayerViewControllerDelegate {
 
     func playerViewController(_ playerViewController: AVPlayerViewController, willBeginFullScreenPresentationWithAnimationCoordinator coordinator: UIViewControllerTransitionCoordinator) {
